backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
pytest
httpx
